Attempted conditional rendering but it didn't work out.
