+++ src/components/MenuList/MenuList.jsx
@@ -1,28 +1,10 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
-import { useDispatch } from 'react-redux';
+import MenuItem from '../MenuItem/MenuItem.jsx';
 
 function MenuList() {
 
-    // Set up Dispatch
-    const dispatch = useDispatch();
-
   const pizzas = useSelector((store) => store.menuReducer);
-  const cart = useSelector((store) => store.cartReducer);
-
-  function addPizzaToCart(pizzaObject) {
-    dispatch({
-      type: 'ADD_PIZZA_TO_CART',
-      payload: pizzaObject,
-    })
-  }
-
-  function removePizzaFromCart (pizzaObject) {
-    dispatch({
-      type: 'REMOVE_PIZZA_FROM_CART',
-      payload: pizzaObject
-    })
-  }
 
 // We need to import the pizza table from Postico and return the pizzas to display:
   // We have to loop through the items. Instead of returning yay a pizza lives here, how can we return the pizzas? 🤔
@@ -32,14 +14,9 @@ function MenuList() {
 
   return (
     <div>
-        {pizzas.map((pizza, i) =>
-          <ul key={i}>
-            <p><img src={pizza.image_path} height="100" width="100"/></p>
-            <p>{pizza.name}</p>
-            <p>{pizza.description}</p>
-            <p>{pizza.price} </p>
-            <button onClick = { () => {addPizzaToCart(pizza)}}>Add to Cart</button>
-            <button onClick = { () => {removePizzaFromCart(pizza)}}>Remove from Cart</button>
+        {pizzas.map((pizza) =>
+          <ul key={pizza.id}>
+            <MenuItem pizzaObject={pizza} />
           </ul>
         )}
     </div>
